chg: ui widget package to be separate of editor
