+++ src/view/show_mode/show_ui_widgets/autotracker/CropTab.py
@@ -4,7 +4,7 @@ from PySide6.QtWidgets import QGridLayout, QLabel, QLayout, QSlider
 
 from controller.autotrack.Helpers.ImageHelper import cv2qim, draw_overlay
 from controller.autotrack.Helpers.InstanceManager import InstanceManager
-from view.show_mode.editor.show_ui_widgets.autotracker.GuiTab import GuiTab
+from view.show_mode.show_ui_widgets.autotracker.GuiTab import GuiTab
 
 
 class CropTab(GuiTab):
