+++ src/view/show_mode/show_ui_widgets/autotracker/GuiTab.py
