+++ src/view/show_mode/show_ui_widgets/autotracker/VFilterLightController.py
@@ -6,7 +6,7 @@ from controller.autotrack.LightController import LightController
 
 if TYPE_CHECKING:
     from model.virtual_filters import AutoTrackerFilter
-    from view.show_mode.editor.show_ui_widgets.autotracker.UIWidget import AutoTrackerUIWidget
+    from view.show_mode.show_ui_widgets.autotracker.UIWidget import AutoTrackerUIWidget
 
 logger = getLogger(__file__)
