+++ src/view/show_mode/show_ui_widgets/pan_tilt_constant_show_ui.py
